+++ CHANGELOG.md
@@ -5,6 +5,7 @@
 ### :wrench: Changed
 
 - Use Python 3.14 in Docker images.
+- Unsupported regex patterns are now removed with a warning instead of causing schema errors.
 
 ## [4.6.6](https://github.com/schemathesis/schemathesis/compare/v4.6.5...v4.6.6) - 2025-12-03